+++ addons/main/script_version.hpp
@@ -1,4 +1,4 @@
 #define MAJOR 2
 #define MINOR 14
-#define PATCH 7
-#define BUILD 1756
+#define PATCH 8
+#define BUILD 1761