+++ addons/sys_cbrn/config.cpp
@@ -1,6 +1,4 @@
 #include "..\main\cfg\MSF_Macros.hpp"
-#include "cfg\MSF_CBRN_Functions.hpp"
-#include "cfg\CBRN_Particles.hpp"
 
 class CfgPatches
 {
@@ -15,6 +13,10 @@ class CfgPatches
         requiredAddons[] = {"3DEN", "cba_settings", "ace_interaction", "MSF3den"};
     };
 };
+
+#include "cfg\MSF_CBRN_Functions.hpp"
+#include "cfg\CBRN_Particles.hpp"
+#include "cfg\ACE_Medical.hpp"
 
 class CfgVehicles
 {